[WIP] WebGPU HAL prototype This is using the webgpu-native C header allowing us to (in theory) compile without needing emscripten for build testing and also will let us use dawn or wgpu in native environments.
This does not use dynamic symbols as they are not yet supported for
WebGPU in emscripten (and may never be a good idea, given some emscripten
limitations). Unlike CUDA and Vulkan I don't forsee us having a native
deployment where WebGPU may be optional so the direct usage is fine for
now.