Make elementwise parallel operations use in-place updates only for non-vectorizable ops.
The `ConvertToDestinationPassingStyle` pass converts elementwise
operations fused with named op producers to use the result of the
named ops as the `outs` operands, i.e.

```
%conv = linalg.conv
%result = linalg.generic ins(%conv, ... : ...) outs(...) -> ...
```

would be converted to

```
%conv = linalg.conv
%result = linalg.generic ins(... : ...) outs(%conv : ) -> ...
```

This avoids using intermediate buffers during bufferization to carry
the dependence. This is only needed when the producer and consumer are
not vectorized since in those cases the dependence is carried through
registers. Make this transformation only when ops dont vectorize.

Depends on https://reviews.llvm.org/D135201